+++ lib/noop_backup/jobs/backup_job.rb
@@ -1,0 +1,9 @@
+module NoopBackup
+  class BackupJob < ActiveJob::Base
+    queue_as :backups
+
+    def perform
+      NoopBackup::Commands::Backup.execute
+    end
+  end
+end